function edits

+++ ashley/tuning properties/vonmisesReliableFit.m
@@ -20,12 +20,12 @@ for iCell = 1:nCells
         y_fit(:,2:nboot+1,iCell) = nan(size(y_fit,1),nboot,1);
     else
     [b(1,iCell), k(1,iCell), R(1,iCell),u(1,iCell),~,R_square(1,iCell)] = ...
-        miaovonmisesfit_ori(deg2rad(theta),data(iCell,:));
+        miaovonmisesfit_ori(deg2rad(double(theta)),data(iCell,:));
     y_fit(:,1,iCell) = b(1,iCell)+R(1,iCell).*exp(k(1,iCell).*(cos(2.*(deg2rad(theta_smooth)-u(1,iCell)))-1));
     [~, max_ori(1,iCell)] = max(y_fit(:,1,iCell),[],1);
     for i = 2:nboot+1
         [b(i,iCell), k(i,iCell), R(i,iCell),u(i,iCell),~,R_square(i,iCell)] = ...
-            miaovonmisesfit_ori(deg2rad(theta),data_resamp(iCell,:,i-1));
+            miaovonmisesfit_ori(deg2rad(double(theta)),data_resamp(iCell,:,i-1));
         y_fit(:,i,iCell) = ...
             b(i,iCell)+R(i,iCell).*exp(k(i,iCell).*...
             (cos(2.*(deg2rad(theta_smooth)-u(i,iCell)))-1));
